refactor: change price type to long

+++ src/main/java/gift/dto/ProductResponseDto.java
@@ -3,7 +3,7 @@ package gift.dto;
 public record ProductResponseDto(
     Long id,
     String name,
-    int price,
+    long price,
     String imageUrl
 ) {
 }
